simulator/requirements.txt
google-genai
python-dotenv
pandas
finnhub-python
schedule
quantstats
yfinance
pydantic
